sanity.bbclass: Run our checking at BuildStarted, not ConfigParsed
ConfigParsed is often utilized to do programmatic changes to the configuration
metadata.  If these other event handlers set any variables which
sanity.bbclass needs to check, there's a problem, since we can't ensure those
handlers run before this one.  So, we move the sanity checking to when the
build is about to start, this ensures that the configuration metadata is
entirely ready to go.

Signed-off-by: Chris Larson <chris_larson@mentor.com>