packaged-staging: Make setting PSTAGING_ACTIVE be a named python func
When the function to set PSTAGING_ACTIVE is an anon python function
overrides are not set so we cannot set things to disabled in the context
of BBCLASSEXTEND recipes.  Call this function before we check
PSTAGING_ACTIVE.  In the case of packagedstaging_fastpath it's cleaner
to convert to a full python function.  While we're in here drop a
duplicate test in populate_sysroot_postamble and make do_package_stage
check PSTAGING_ACTIVE for == 0 rather than != 1, like the rest of the
functions do.

Signed-off-by: Tom Rini <tom_rini@mentor.com>