packaged-staging, native{,sdk}: Rework OVERRIDES to check PSTAGING_DISABLED
When the function to set PSTAGING_ACTIVE is an anon python function OVERRIDES
is not set so we cannot set things to disabled in the context of BBCLASSEXTEND
recipes.  To fix this we need to have native/nativesdk set OVERRIDES outside of
the anon function context.  We then make a COW of the data at this point in
packaged-staging's anon function and evaluate so we know what will be in
OVERRIDES and check.  While we're in here drop a duplicate test in
populate_sysroot_postamble and check nativesdk along with the rest of the
classes that need mangling.

Signed-off-by: Tom Rini <tom_rini@mentor.com>