cmake-2.8.2: Add recipe for non-native cmake 2.8.2.
Tested on om-gta02:
* builds trivial C and C++ executable
* "include(FindPkgConfig)" correctly finds the cmake module
* pkg_check_modules correctly invokes pkg-config
* ccmake works

Also tested on babbage by Eduardo Lima (thanks for help).

Unfortunately, useful doc is not packaged. The reason is that the doc
is produced by running the built binaries. Changing this would probably
require a complex patch of the build process. Luckily, the doc is
still contained in the original binaries, so it's accessible on target
by e.g. cmake --help-full.

Signed-off-by: David Kozub <zub@linux.fjfi.cvut.cz>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>