libprelude: new recipe
Libprelude is a library that guarantees secure connections between
all sensors and the Prelude Manager. Libprelude provides an Application
Programming Interface (API) for the communication with Prelude
sub-systems, it supplies the necessary functionality for generating and
emitting IDMEF events with Prelude and automates the saving and
re-transmission of data in times of temporary interruption of one of
the components of the system.

Signed-off-by: Roman I Khimov <khimov@altell.ru>