ncurses: added 5.7 recipe
It uses ideas from the old 5.4 recipe but brings a lot of new features
like widec libs, pkgconfig files or splitted library packages.

As lot of packages depend on ncurses, whole distribution must be
probably rebuild when updating to this new version.

Due to this and the complex recipe, it has a negative DEFAULT_PREFERENCE
for now.

Signed-off-by: Enrico Scholz <enrico.scholz@sigma-chemnitz.de>