u-boot: beagleboard: default 3MP camera support
This turns on the pull-ups so that the I2C will work by default as well
as sets the default camera type.  Currently, the kernel cannot probe for
I2C devices, but that should be the long term fix.  We put the change
here into the default environment, instead of the boot.scr, such that
there won't be a need to update your boot.scr in the future for the
default camera if we ever get the kernel fixed to probe for I2C devcies.

Signed-off-by: Jason Kridner <jkridner@beagleboard.org>
Signed-off-by: Koen Kooi <koen@openembedded.org>