tofrodos-native: created
Created from the same version used for non-native.

Had to replace S, since PN doesn't reflect the tar directory (tofrodos).

Updated patch to include creation of new directories.  This needed to be
done in the Makefile to cover both do_install and do_populate_sysroot.

Added patch to use environment variables, rather than passed in paths.

This could probably be done by enabling native builds in the main
recipe, but I couldn't figure out how to do that with some of the extra
stuff in it.  I'd suggest accepting this patch until there is one to
replace it.

Signed-off-by: Jason Kridner <jkridner@beagleboard.org>
Signed-off-by: Eric Bénard <eric@eukrea.com>