base.bbclass: fix bug with absolute file:// uris
The code which used host/path from the url to construct the destination of the
copy was using os.path.join.  os.path.join treats a "/" in an element as an
indication to drop all previous elements, so the full path resulted in a
destination exactly the same as the source, resulting in a no-op unpack.
Fixed by using oe.path.join instead.

Signed-off-by: Chris Larson <chris_larson@mentor.com>