ncurses: use linker scripts for libncurses(w)
Some software (e.g. util-linux-ng) assumes that symbols from -ltinfo
will be added when it is linked against -lncurses.  This breaks when
linkerflags are containing --no-copy-dt-needed-entries which is the case
e.g. in Fedora 13+.

This patch replaces the libncurses.so symlink with a linkerscript which
adds -ltinfo.  To do things in a consistant manner, the patch rewrites
most parts of 9e58422868a03eeafa72781477174850d1c5abe7.

NOTE: it might require a 'bitbake ncurses -c clean' to see an effect
because staging on updates does not change a symlink (libncurses.so) to a
regular file.

Signed-off-by: Enrico Scholz <enrico.scholz@sigma-chemnitz.de>