pulseaudio: made 0.9.21 compile for armv4 and armv5 too
there was some armv6 asm in here.
Fortunately the function with much inline asm was only
called for armv6 or higher so this is ifdef'd out for all
armv4 and armv5 variants.
Also there was a single instruction but the code also had a C
alternative so again only some defines were needed

Signed-off-by: Frans Meulenbroeks <fransmeulenbroeks@gmail.com>