Elmdentica abandoned json-c in favor of cjson
Hi,

cJSON is not a properly setup library, but drop in files (this should be
fixed in the future), however it seems to be working a lot better than
json-c.

I was running into a lot of weird problems and crashes derived from
json-c and so I tried cJSON as advised by disconfitor.

Indeed it's working, faster, and more reliably, so I now add two patches:
  1) removes elmdentica svn dependency on libjson (already commit in e's
svn)
  2) removes my naive patch to json-c (or libjson) as they're working on
a much better patch and I was the only one using it AFAICT

Please apply!

Rui

Signed-off-by: Klaus Kurzmann <mok@fluxnetz.de>