kernel.bbclass: Fix INITRAMFS_IMAGE logic
Due to a bitbake bug to be fixed in 1.10.1, EXPORT_FUNCTIONS isn't
remapping kernel_do_compile[depends] to do_compile[depends], so
make a comment about cleaning this up once we depend on a working
bitbake.  Next, it needs to add to do_configure (like the users
have manually set a dependancy on) rather than do_compile so that
the initramfs is copied into place in time.  Since we're fixing the
logic in kernel.bbclass, we can drop the workarounds that linux-kexecboot
and linux-preboot have added.

Signed-off-by: Tom Rini <tom_rini@mentor.com>