wesnoth: add inc factorized 1.8.4 version, with normal and wvga versions
Wesnoth is a strategy game and features more than one type of GUI,
each GUI type(normal,smallGUI,tinyGUI) has some minimum resolution constraint.
If the minimum resolution is not met, the game abort due to some asertions
in the code.

Unlike the choice between normal and smallGUI that can be made at runtime,
choosing,or not the --smallgui comand line option, the tinyGUI choice must
be defined at compile time(configure option),because it scales the graphics.

On devices like the nokia900, which have a 800x480 resolution,
the tinyGUI is not necessary anymore, and it's better to disable it because
the tinyGUI has some scalling issues among other GUI related issues.
(vilagges can only be distinguished as a flag)

The wvga version is like the Desktop version of wesnoth,if someone wants
to play it on a wvga screen he has to run wesnoth --smallgui.

Note that wesnothd was renamed to wesnoth-server(conversation from #oe on freenode):
  <GNUtoo|laptop> I've locally made wesnoth into an inc and made wesnoth-wvga and wesnoth
  <GNUtoo|laptop> the issue is wesnothd
  <GNUtoo|laptop> I rename it to ${PN}-server
  <GNUtoo|laptop> ?
  <GNUtoo|laptop> what about the upgrade path
  * dth_ntb (~dth@a89-183-26-221.net-htp.de) has joined #oe
  <woglinde> sure why not
  <woglinde> you are the only active user so far
  <woglinde> so go on