utility-tasks.bbclass: drop mrproper task
This task is misleading and potentially quite harmful, as it wipes the entire
DL_DIR, yet tasks are supposed to be recipe specific.  We have the proper
distclean task to wipe sources for a recipe + run do_clean, and the user can
always wipe all of DL_DIR manually if necessary.

Signed-off-by: Chris Larson <chris_larson@mentor.com>
Acked-by: Holger Hans Peter Freyther <holger@moiji-mobile.com>
Acked-by: Eric Bénard <eric@eukrea.com>