nostromo: add new recipe for this web server
* description taken from its website :
nhttpd is a simple, fast and secure HTTP server.
It runs as a single process, handling connections with select(2).
For CGIs and directory listing it does fork(2). nhttpd has the
minimum of HTTP/1.1 and CGI/1.1 implemented. Also supported are;
chroot, setuid, basic authentication, SSL, IPv6, custom repsonses,
aliases, and virtual hosts. To stop the server send a SIGTERM singal
to the PID. The access log is written in standard CLF format.

Signed-off-by: Eric Bénard <eric@eukrea.com>