linux-libc-headers_2.6.24: Add falloc.h to kernel exported headers
This adds include/linux/falloc.h to the list of files exported by the kernel to the C library.
The include file is required for building e.g. a recent util-linux-ng against 2.6.24.

Starting from 2.6.25 a similar patch has been accepted upstream (the mainline
kernel). For 2.6.24, however, falloc.h is not exported although it's present.

Signed-off-by: Thilo Fromm <t.fromm@dresearch.de>
Signed-off-by: Khem Raj <raj.khem@gmail.com>