python-{native}-2.6.5: Fix parallel build.
Randomly the build failed for me with the following error message.

    libpython2.6.so: undefined reference to `_PyParser_Grammar'

The applied patch in [5] tried to fix this problem, but it turns out that this also was due to a possible race. This is already known in OpenOffice too [3].

R. David Murray (dmalcolm) <dmalcolm@fedoraproject.org> fixed this in Fedora [1] and I applied his patch.

I reported this issue upstream [4] and this will only be fixed in Python 2.7, so we have to keep this patch for Python 2.6.

I put 00 in front of the name of the patch file because I did not want to rename the other patches.

I build tested this with `MACHINE = " beagleboard"` for `angstrom-2008.1` and `minimal-{eglibc,uclibc,libc}`.

[1] http://pkgs.fedoraproject.org/gitweb/?p=python.git;a=commitdiff;h=b95f6cc2ca6a009f97436c6aa16cfd70547353d9
[2] http://www.openoffice.org/issues/show_bug.cgi?id=114866
[3] http://doc.services.openoffice.org/wiki/RedTinderboxStatusInEIS
[4] http://bugs.python.org/issue10013
[5] http://patchwork.openembedded.org/patch/1899/

Signed-off-by: Paul Menzel <paulepanter@users.sourceforge.net>
Signed-off-by: Khem Raj <raj.khem@gmail.com>