perl: updated LICENSE; GPL -> GPLv1
perl has as license the choice between Artistic or GPLv1.
see http://dev.perl.org/licenses/
so changed LICENSE = "Artistic|GPL" to LICENSE = "Artistic|GPLv1"

This also applies to CPAN recipes. Typically they specify that the
code is licensed under the same terms as perl so figured that the
above changes would be better in most (if not all) cases
(and no, I did not verify all 130 or so recipes, if you feel this is needed,
be my guest)

Signed-off-by: Frans Meulenbroeks <fransmeulenbroeks@gmail.com>