bitbake.conf: fix pkg-config relocation issue
By default, pkg-config falls back to checking the default search path for .pc
files, which is the native location.  If you relocate it, it will be unable to
hit that location, which can cause problems in certain cases.  In particular,
gnome-bluetooth requires gtk-doc-native to run its tools, and uses pkg-config
to determine its availability.

Fix by adding the native pkg-config locations to the PKG_CONFIG_PATH.

Signed-off-by: Chris Larson <chris_larson@mentor.com>
Acked-by: Khem Raj <raj.khem@gmail.com>