distro, toolchain-external: override TARGET_VENDOR with TOOLCHAIN_VENDOR
Trying to use an external toolchain like the 2009q1 toolchain from codesourcery
has some strange results since the toolchain has vendor 'none' but
distributions like angstrom set TARGET_VENDOR unconditionally.

This patch overrides the TARGET_VENDOR with the value of TOOLCHAIN_VENDOR,
if set. TOOLCHAIN_VENDOR can be set in the local configuration alongside the
other TOOLCHAIN_* variables. The override is performed in
toolchain-external.inc, as suggested by Koen Kooi [1]. It also reorders the
inclusion of the distro and toolchain-* so that toolchain-external can override
the value of TARGET_VENDOR, as suggested by Denys Dmytriyenko [2].

Tested with angstrom-2008.1 and the codesourcery 2009q1 toolchain by assigning
TOOLCHAIN_VENDOR = "-none"
TOOLCHAIN_TYPE = "external"
TOOLCHAIN_BRAND = "csl"
in addition to putting the CSL in the PATH.

[1] http://thread.gmane.org/gmane.comp.handhelds.openembedded/37918/focus=38047
[2] http://thread.gmane.org/gmane.comp.handhelds.openembedded/38176/focus=38188

Signed-off-by: Ben Gardiner <bengardiner@nanometrics.ca>
CC: Koen Kooi <k.kooi@student.utwente.nl>
CC: Víctor M. Jáquez L. <vjaquez@igalia.com>
CC: CC: Denys Dmytriyenko <denis@denix.org>
Signed-off-by: Koen Kooi <koen@openembedded.org>
Signed-off-by: Denys Dmytriyenko <denis@denix.org>