libgee: define common INC_PR and use it in all specific version recipes
Until now there is now way to indicate that the libgee.inc common part of libgee has
changed - the PR is only defined in the specific version recipes. This adds the INC_PR var
to to the common libgee.inc and use it in the specific version recipes.
INC_PR is initial set to the highest value of both version 0.5.2 and 0.6.0.

Signed-off-by: Simon Busch <morphis@gravedo.de>
Signed-off-by: Khem Raj <raj.khem@gmail.com>