gdb-cross/cross-sdk: Fixup DEPENDS further
Since we are statically linking, as noted before, we now share
the same DEPENDS set on cross and cross-sdk.  And gdb cannot
build without grabbing some stuff from the real compiler.
gdb-cross was fine but gdb-cross-sdk was relying previously on
an oddity of zlib.  To be clear, add in virtual/${TARGET_PREFIX}gcc
with a comment.

Signed-off-by: Tom Rini <tom_rini@mentor.com>