gdb-cross-sdk.inc: We must use do_configure_append
The goal is to have oe_runconf happen twice.  So we want
the normal gdb-cross.inc do_configure to happen so we get
an initial configure.  Then gdb-cross-sdk does its thing and
then we run oe_runconf again.

All of this shows that we really need to sit down and think and
test the dance we're doing here.

Signed-off-by: Tom Rini <tom_rini@mentor.com>