Reverse the order of OVERRIDES
Given the current implementation of OVERRIDES in bitbake, the variable is
expected to contain elements in the order least specific to most specific,
however, our current usage of it does not match that.  As one example, "local"
is supposed to always be the most specific override, yet currently it's the
least specific.  As another example, currently the target architecture is seen
as more specific than the machine, which is also clearly wrong.

Big thanks to Chase Maupin for investigating and identifying this long
standing issue.

It becomes clear that a reversal of the current value will bring us to a more
sane behavior, and avoids the need for the dual overrides hack mentioned in
the comments, so this implements this reversal, and drops the unnecessary and
confusing comments.

This also introduces a MACHINE_OVERRIDES variable as a generic mechanism to
inject overrides elements which are more specific than the distro but less
specific than the machine, which is where things like MACHINE_CLASS or
SOC_FAMILY or the like would go.  This variable is *space* separated, to make
it easier and more convenient to assemble the variable incrementally.

Reported-by: Chase Maupin <chase.maupin@ti.com>
Signed-off-by: Chris Larson <chris_larson@mentor.com>
Acked-by: Chase Maupin <chase.maupin@ti.com>
Acked-by: Khem Raj <raj.khem@gmail.com>