autotools.bbclass: Specify a value of sysroot instead of asking the compiler
If we just use --with-libtool-sysroot then it asks the compiler at
configure time to supply the sysroot which will not work with external
toolchains because they may have different default sysroot encoded into
them. Therefore we specify the sysroot so the configure can use that
value and not depend upon compiler to supply one.

We do not do it for native recipes as the native compiler should
fall back to /usr prefix and not solely depend on sysroot otherwise
we end up staging everything in native sysroot before we start to build
target recipes

Signed-off-by: Khem Raj <raj.khem@gmail.com>