metadata_scm: use rev-parse rather than show-ref
show-ref will show all matching refs, so a "show-ref HEAD" will show not just
the local HEAD, but the fetched remote ones as well.  This isn't what we want
for this function, so change it to use rev-parse with --verify, and also
change it to use --short, to shorten the hash to a more palatable length.

Signed-off-by: Chris Larson <chris_larson@mentor.com>
Acked-by: Tim Harvey <harvey.tim@gmail.com>
Acked-by: Khem Raj <raj.khem@gmail.com>