tune-mips32.inc: Add missing definitions.
I added the FEED_ARCH and BASE_PACKAGE_ARCH definitions which
are assigned the value of TARGET_ARCH (which is defined in
the individual machine files). I used the TARGET_ARCH variable
and not a static value since the value can be either "mipsel"
(as in ben-nanonote) or "mips" (as in qemumips).

Signed-off-by: Alex Ferguson <thoughtmonster@gmail.com>
Signed-off-by: Kristoffer Ericson <kristoffer.ericson@gmail.com>