libsdl-x11 1.2.14: removing libtool-update.patch fixes alsa detection
A side effect of the patch is that alsa detection was disabled or didn't work.

For instance here's a short way to test the result of ./configure:

With the patch:
grep ALSA SDL-1.2.14/include/SDL_config.h gives:
  /* #undef SDL_AUDIO_DRIVER_ALSA */
  /* #undef SDL_AUDIO_DRIVER_ALSA_DYNAMIC */

And without the patch:
grep ALSA SDL-1.2.14/include/SDL_config.h gives:
  #define SDL_AUDIO_DRIVER_ALSA 1
  #define SDL_AUDIO_DRIVER_ALSA_DYNAMIC "libasound.so.2"

Note that libsdl-x11 builds fine without the patch under
  the SHR distribution with the htcdream and nokia900 machines

Thanks demonicmaniac for help in #sdl on Freenode.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@no-log.org>
Acked-by: Khem Raj <raj.khem@gmail.com>