Revert "gcc-cross4.inc: We dont support multilib so empty out the with-libdir-list for SH4"
This fixed building the toolchain though but broke
kernel as it expect gcc to be able to compile no-fpu
code. Which means it will break all gcc 4.x

So we revert it and keep the status quo of 4.5 being broken
but older 4.x gcc can still build for SH based machines.

This reverts commit 35881fcc6feb9670947056ff868e03b0a0900a51.