help2man: Add native recipe, drop check for being provided.
With a recent change to auto* to not build the man pages for
native recipes, we can make build progress without help2man
being there already.  This adds help2man to local.conf.sample
as an example of something that is assumed to be provided,
drops the sanity.bbclass check and adds a help2man-native recipe.
In order to catch the cases where help2man is likely to have been
an implicit requirement before we add it to the list that
autotools brings in.

Signed-off-by: Tom Rini <tom_rini@mentor.com>
Acked-by: Michael Smith <msmith@cbnco.com>