netsurf: Update netsurf, dependancies, add netsurf-fb recipe.
Updated netsurf-2.1 to netsurf-2.6 and updated fix_makefile.patch and
debugxml_fix.patch for netsurf-2.6.

Updated hubbub, libnsbmp, libnsgif, libparserutils from version 0.0.1
to version 0.0.2.

Added two new recipes libcss and libwapcaplet which are needed by
netsurf-2.6. Recipe libcss comes with patch no-werror.patch which
removed the definition of -Werror from the makefile for building
with gcc-4.5, as per the recommendation of the netsurf developers.

Added netsurf-fb recipe, which builds netsurf with a framebuffer
frontend, which is much lighter than the GTK frontend (and can run
under a plain tty). The framebuffer frontend is set to use SDL because
it's the most stable backend at the moment. Font rendering is handled
by freetype and uses the DejaVu fonts.

Added libnsfb, a framebuffer abstraction layer library needed by
netsurf-fb along with a patch, no-werror.patch, which removes the
definition of -Werror from the makefile for building with gcc-4.5.

Signed-off-by: Alex Ferguson <thoughtmonster@gmail.com>
Signed-off-by: Kristoffer Ericson <kristoffer.ericson@gmail.com>