update-rc.d: make update-rc.d local
the recipe did not build any more as handhelds.org is gone.
Since this one is fairly crucial as lots of recipes depend on it
I decided to put the update-rc.d file in update-rc.d/ instead of
fetching it.

It has been decided (after discussion on irc with Philip Balister, the maintainer of
update-rc.d) to keep the code the same as the old upstream
version and not apply the patch to it.

While at it als renamed the files dir to update-rc.d

Signed-off-by: Frans Meulenbroeks <fransmeulenbroeks@gmail.com>
Acked-by: Philip Balister <philip@balister.org> (on irc)