ti-pru-sw-app-loader: first addition of PRU sw application loader recipe
* Add PRU sw application loader recipe for the omapl138 machines

This recipe builds the PRU application loader that is used to load a
binary to the PRU's memory and manage the code executed in the PRU from
the user space.  The library and include files from the application
loader are staged for use by the ti-pru-sw-examples recipe.  Details
about the application loader can be found at:
http://processors.wiki.ti.com/index.php/PRU_Linux_Application_Loader

Signed-off-by: Melissa Watkins <m-watkins@ti.com>
Signed-off-by: Chase Maupin <chase.maupin@ti.com>
Signed-off-by: Koen Kooi <k-kooi@ti.com>