gob2: updated to 2.0.17
When trying to build gmpc on our autobuilder (64 bit ubuntu 10.04) I got compile errors
because gob2 apparently corrupted some data it generated causing type names to be changed.
On opensuse 11.2, 32bit it worked like a charm.
As our gob2 was 2.0.14 from 2006 and the latest one was 2.0.17 from earlier this year I
decided to update first.
Apparently this has fixed the issue. (32 and 64 bit work fine)
Also merged the native recipe in the target recipe.

Signed-off-by: Frans Meulenbroeks <fransmeulenbroeks@gmail.com>