gcc-4.5.inc/gcc-svn.inc: Dont do the sourcedir move dance.
In some cases the rename_srcdir ended up moving the sourcedir
as a sub-directory under gcc-4.5 which was not intended actually
it is intended to be renamed and not moved. Which was seen recently
where people complained about a patch not being applied to gcc-4.5
even though patch was not changed. The reason was the srcdir was
moved instead of renaming. This could happen if target dir already
existed.

Anyway so we get rid of this dance completely. So nip the evil
in the bud

Signed-off-by: Khem Raj <raj.khem@gmail.com>