ust.inc: Rework FILES
libust.so is actually an ld script (grouping libust.so.0 and
libust_initializer.o), so I doubt it has any usefullness on the target.  The
note about preloading should concern other libust*.so, that are indeed used in
preloading by 'usttrace' script. And these .so libraries are going to target,
not to -dev. And for libust.so that script preloads libust.so.0.

Signed-off-by: Andrew Gabbasov <andrew_gabbasov@mentor.com>
Signed-off-by: Tom Rini <tom_rini@mentor.com>