util-linux-ng: Clean up native config bits.
We don't need the ncurses stuff for how we use this, so disable it
(and bump INC_PR).  In util-linux-ng_2.17.bb the override to add
some bits to the native recipe was killing all of the other parts
so change the syntax so it doesn't, and set PR=INC_PR (now that
INC_PR is bumped).

Signed-off-by: Tom Rini <tom_rini@mentor.com>