perl: blacklist perl-module-cpanplus-internals-source-sqlite for perl-modules
It has runtime dependencies on libdbd-sqlite-perl and libdbix-simple-perl
which doesn't get built automatically if not required as build time dependency
for something.

It's not possible to add them as build time dependencies for
perl itself since they're in turn depend on perl making a circular dependency
(and they _really_ depend on perl because they build binaries, so no way to
cheat by making this recipes not to depend on perl).

It's also not correct to downgrade this runtime dependency to recommendation,
since the module is really unusable without its dependencies.

The most annoying thing about all this is that this dependency situation renders
perl-modules package uninstallable by itself, failing with something like:

 Collected errors:
  * satisfy_dependencies_for: Cannot satisfy the following dependencies for
 task-sdk-native:
  *    libdbd-sqlite-perl *    libdbix-simple-perl *

So the best way to solve it IMO is to just blacklist it from perl-modules
since the value of this module is questionable. It's experimental CPANPLUS
backend that's not used by default, you have to specifically configure your
perl to use it. And if you _really_ want to use it - the module package itself
will still contain proper dependencies, you just need to get them built and
installed.

The problem only exists in perl 5.10.1, but 5.8.8 doesn't have this module, so
blacklisting in common .inc doesn't hurt.

Signed-off-by: Roman I Khimov <khimov@altell.ru>
Acked-by: Khem Raj <raj.khem@gmail.com>