busybox: move syslog config to /etc/default
The busybox syslog syslog.conf is parsed by the /etc/init.d script, not by the
syslog process itself, so it belongs in /etc/default.  In addition, the file
format is *completely* different from the standard sysklogd configuration, so
while we should resolve the file conflict between busybox-syslog and sysklogd,
we should not use update-alternatives for it, so this is a cleaner solution.

Note: when upgrading with a modified /etc/syslog.conf, the old config will
stick around as /etc/syslog.conf-opkg.backup, and the user is then free to
move this to the new /etc/default location.  This could be automated, but we
can't risk moving a real syslog /etc/syslog.conf into the busybox config file
location, so this is best.

Signed-off-by: Chris Larson <chris_larson@mentor.com>
Acked-by: Koen Kooi <koen@openembedded.org>
Acked-by: Khem Raj <raj.khem@gmail.com>