u-boot: Make sure we use override HOSTCC for env tools
Upstream says that the right way to build a target binary of the env
tools is to pass in HOSTCC=your-cross-cc.  Bump the PR in u-boot_git.bb
to fix as many easily fixable ones as possible, but other platforms
may have been shipping bad binaries (since we had skipped testing
on this package due to GNU_HASH issues when built for the target).

Signed-off-by: Tom Rini <tom_rini@mentor.com>