native: fix duplicating virtclass-native entries in OVERRIDES
A function was called from ${@} in OVERRIDES which was supposed to return the
bits to prepend when in virtclass context, yet it *set* overrides instead,
returning None.  This resulted in 1) adding an extra virtclass-native to
OVERRIDES each time it was expanded, and 2) appending None, causing
'localNone' to be in overrides rather than the expected 'local'.

Signed-off-by: Chris Larson <chris_larson@mentor.com>