lzma: Various cleanups and a bugfix
We use the subdir parameter to SRC_URI so that we will unpack into
the normal S directory rather than WORKDIR (which avoids a race with
the append to unpack and removal of empty log files).  Next, we replace
the regex to override the compiler with a patch that really does it
where it's needed (C/LzmaUtil) and now we have a target liblzma.a
rather than host liblzma.a.  For CPP/7zip/Compress/LZMA_Alone we
just need to pass in correct values via CXX_C / CXX.  Given
the structure of the Makefile it's easiest to pass in the rest
of the flags as part of the compiler.

Signed-off-by: Tom Rini <tom_rini@mentor.com>