groff: Fix do_install in certain cases, along with DEPENDS
Sometimes groff fails in the do_install step, complaining that img/pic* does
not exist.  Normally, these files would be created by groff itself during the
build process, but since we are cross-compiling, the old recipe replaces the
invocation of groff and troff with echo commands.  However, echo does not
create the output files, which eventually results in the do_install failure.

This patch makes groff dependent on groff-native, and instead of using echo,
it instructs the build process to use the staged groff binaries.

(Tom Rini: While in here, switch to GNU_MIRROR)

Acked-by: Khem Raj <raj.khem@gmail.com>
Signed-off-by: Mike Westerhof <mike@mwester.net>
Signed-off-by: Tom Rini <tom_rini@mentor.com>