opkg-utils: convert mtime to int before comparing it
The st_mtime attribute (which is a float) is compared against a value
from the timestamp database, which was stored as an integer there.

When working on a filesystem with precise timestamps the comparision
will fail nearly everytime hence.

Although it might be possible to enhance the database to store the
fractional part too, this will complicate things more than we would
gain by this change.

This patch was applied as 49e11b32906fc1becd9218bb3125f6fc8fa03fa4 to
old ipkg-utils already.

Signed-off-by: Enrico Scholz <enrico.scholz@sigma-chemnitz.de>
Acked-By: Graham Gower <graham.gower@gmail.com>
Signed-off-by: Khem Raj <raj.khem@gmail.com>