package.bbclass: Update locking now that pstaging is always on.
In a few functions we checked if pstagingactive was set and if so, grabbed
the lock.  But really, we should always be locking here.  And inside of
package_do_shlibs we need to keep the lock a little longer than we had
been otherwise we can run into a race..

Signed-off-by: Tom Rini <tom_rini@mentor.com>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>