xfce_session_4.6.{1,2}: explicitely pass the path of `iceauth`
Xfce Session build depends on `iceauth`, which fails when cross compiling for a different architecture.

As suggested by Khem Raj [1] pass the value, `${bindir}/iceauth` is taken as suggested by Koen Kooi [2], to the configure script so that the test is avoided.

Therefore `iceauth` is removed from `DEPENDS` and the PR variable is incremented.

This issue has been reported upstream as ticket #7420 [3].

[1] http://lists.linuxtogo.org/pipermail/openembedded-devel/2011-March/031115.html
[2] http://lists.linuxtogo.org/pipermail/openembedded-devel/2011-March/031188.html
[3] http://bugzilla.xfce.org/show_bug.cgi?id=7420

Signed-off-by: Paul Menzel <paulepanter@users.sourceforge.net>
Acked-by: Koen Kooi <koen@dominion.thruhere.net>
Acked-by: Khem Raj <raj.khem@gmail.com>
Acked-by: Andreas Müller <schnitzeltony@gmx.de>