abuse: update to 0.7.1,fix license,compilation(host libdir),and sdl install
Without the sdl install fix() we have:
| install: cannot stat `.../work/armv7a-angstrom-linux-gnueabi/abuse-0.7.0-r1/abuse.png': No such file or directory

Without the patch -L/usr/lib was passed to the linking flags and it failed at linking.
And if we didn't set the xorg library path we had -L alone in the linking flags which made it fail
  too

Finally The license is assumed to be GPLv2,becase of COPYING.
Altough no headers were present in the source files.

Note that no data are included yet(it could be included in some
  future improvements)

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@no-log.org>