python-numpy: add missing runtime dependencies
Without these runtime dependencies there are errors on the target when executing `mypaint`, because `python-numpy` code calls `import` from the distinct packages.

TODO: It is not clear where this error originates from and what the best fix is. There are the following alternatives.

1. Add the runtime dependencies to `mypaint`.
2. Add these packages to `RRECOMMENDS`.

This patch is build tested using `angstrom-2010.x` for `MACHINE = "beagleboard"`.

Signed-off-by: Christian Charreyre <christian.charreyre@cioinfoindus.fr>
Signed-off-by: Paul Menzel <paulepanter@users.sourceforge.net>
Signed-off-by: Koen Kooi <koen@dominion.thruhere.net>