lmbench: Rework for binaries in /usr/bin
With the move to having lmbench binaries reside in /usr/bin both the
lmbench-run script we provide as well as config-run and results from
lmbench no longer work, so we need to fixup the paths in a few places.
While in here, call install -d once to make all the directories we need
and re-order stuff to flow better.

Signed-off-by: Tom Rini <tom_rini@mentor.com>