task-base: make it possible for a distro to set it's alsa-state provider
There different recipes in OE which provides alsa-state. In my case it's alsa-state and
fsoaudiod which conflicts as both are mostly the same. For this I added a
DISTRO_ALSA_STATE variable which each distro can set on it's own the provider of
alsa-state it want to use. The default provider is the alsa-state recipe.

Signed-off-by: Simon Busch <morphis@gravedo.de>
Acked-by: Martin Jansa <Martin.Jansa@gmail.com>